// src/api/COD/call-of-duty.d.ts
declare enum platforms {
  All = 'all',
  Activision = 'acti',
  Battlenet = 'battle',
  PSN = 'psn',
  Steam = 'steam',
  Uno = 'uno',
  XBOX = 'xbl',
  NULL = '_',
}
declare enum games {
  ModernWarfare = 'mw',
  ModernWarfare2 = 'mw2',
  Vanguard = 'vg',
  ColdWar = 'cw',
  NULL = '_',
}
declare enum friendActions {
  Invite = 'invite',
  Uninvite = 'uninvite',
  Remove = 'remove',
  Block = 'block',
  Unblock = 'unblock',
}
// ...
declare class USER {
  friendFeed: (gamertag: string, platform: platforms) => Promise<any>;
  eventFeed: () => Promise<any>;
  loggedInIdentities: () => Promise<any>;
  codPoints: (gamertag: string, platform: platforms) => Promise<any>;
  connectedAccounts: (gamertag: string, platform: platforms) => Promise<any>;
  settings: (gamertag: string, platform: platforms) => Promise<any>;
  friendAction: (
    gamertag: string,
    platform: platforms,
    action: friendActions
  ) => Promise<any>;
}
declare class ALT {
  search: (gamertag: string, platform: platforms) => Promise<any>;
  cleanWeapon: (weapon: string) => Promise<string>;
}
declare const Warzone: WZ;
declare const ModernWarfare: MW;
declare const ModernWarfare2: MW2;
declare const Warzone2: WZ2;
declare const ColdWar: CW;
declare const Vanguard: VG;
declare const Store: SHOP;
declare const Me: USER;